# backend-apigate/internal/handler/auth.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/hiroki706/microarch-tutorials/backend-apigate/api"
)

func (s *Server) RegisterUser(w http.ResponseWriter, r *http.Request) {
	var req api.UserRegisterRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// ユーザー登録のロジックを呼び出すだけ
	err := s.authUC.RegisterUser(r.Context(), req)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusNoContent) // 204 No Content
}
func (s *Server) LoginUser(w http.ResponseWriter, r *http.Request) {
	var req api.UserLoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	tokenPair, err := s.authUC.LoginUser(r.Context(), req)
	if err != nil {
		http.Error(w, err.Error(), http.StatusUnauthorized)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK) // 200 OK
	json.NewEncoder(w).Encode(tokenPair)
}

func (s *Server) RefreshToken(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "Not implemented", http.StatusNotImplemented)
}